FrameworkTraining/sandbox/requirements.txt
Django==3.0.8
django-debug-toolbar==2.2
django-pytest==0.2.0
djangorestframework==3.11.0
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
requests==2.24.0
django-filter
markdown
